fix: 修复了部分bug

+++ common/src/main/java/com/gaokao/common/meta/po/UserForm.java
@@ -26,7 +26,7 @@ public class UserForm {
 
     private String name;
 
-    private Long score;
+    private Integer score;
 
     private String subject;
